From Martin Kraemer:
When using the "Mark" and "Cursor" positions is emacs mode (set
using ^@, show using ^X^X), there are still some regressions:

If a word is deleted and the Mark is inside the deleted word, then
after deletion the Mark position is wrong.  The appended patch fixes
that.

Enter:
  tcsh> ABC DEF^@123GHI^B^B^B^[d^X^X
Explanation:
  The line contains the text "ABC DEF123GHI".
  The mark is at "1", the cursor at "G".
  A "delete-prev-word" is executed, and the cursor
  is now at the start of the former "DEF..." (now "GHI").
  The whole text has been collapsed to "ABC GHI".
  The ^X^X (exchange-point-and-mark) shows that
  the Mark now is not at the beginning of the
  deleted word, but at "B" in the 1st word.

The patch tests whether the mark is between the cursor and the other
end of the deleted range. If yes, it is set to the same position as the
cursor.