2. $histchars
My second concern is about HISTSUB, as this doesn't seem to like to be
set to a null character, either.  Please look at the following example.

	> set histchars='\!'
	> echo weird
	> !e
	Bad substitute.
	> _

The shell doesn't allow you to use '!' even though HIST is set to '!'.
The error message "Bad substitute" is displayed when you have !!:s//
or something similar in the input and there's something wrong with it.
But clearly there is no such input here.

What is happening here is that the getC1() routine in sh.lex.c calls
getexcl() with 0 for every '!' it encounters in the input, but getexcl()
basically mistakes this for HISTSUB (== 0) and tries to proceed as if
this were a case of !!:s//, which it clearly isn't, leading to the
above-mentioned error message.

My initial patch assigned a space character to HISTSUB when only one
character was set to $histchars.  I chose a space for this because
1) it didn't seem to interfere with the rest of the parsing code, and
2) the parser just skips spaces in the beginning of the input line, so
history substitution by way of HISTSUB would have no chance of being
activated, i.e., it would be effectively disabled, as seemed to be
intended with the line: set histchars='!'.  (I knew this was a bit
tricky but I wanted to modify as few files as possible.)

I realize now there is a better solution to this: allow HISTSUB to be
set to a null character and modify sh.lex.c about its treatment of
HISTSUB.  Here's a patch for this.

=============================================================== >8 ==
--- sh.lex.c.ORG	2010-05-13 00:10:48.000000000 +0900
+++ sh.lex.c	2010-05-16 15:37:49.000000000 +0900
@@ -672,11 +672,11 @@
     eChar c;

     if (sc == 0) {
-	sc = getC(0);
-	if (sc != '{') {
-	    ungetC(sc);
-	    sc = 0;
-	}
+	c = getC(0);
+	if (c == '{')
+	    sc = (Char) c;
+	else
+	    ungetC(c);
     }
     quesarg = -1;

@@ -693,7 +693,7 @@
 	for (ip = hp->next->next; ip != hp->prev; ip = ip->next)
 	    dol++;
     left = 0, right = dol;
-    if (sc == HISTSUB) {
+    if (sc == HISTSUB && HISTSUB != '\0') {
 	ungetC('s'), unreadc(HISTSUB), c = ':';
 	goto subst;
     }
@@ -726,7 +726,7 @@
     exclc = right - left + 1;
     while (--left >= 0)
 	hp = hp->next;
-    if (sc == HISTSUB || c == ':') {
+    if ((sc == HISTSUB && HISTSUB != '\0') || c == ':') {
 	do {
 	    hp = getsub(hp);
 	    c = getC(0);
@@ -1147,7 +1147,7 @@
     int     event;
     int    back = 0;

-    c = sc == HISTSUB ? (eChar)HIST : getC(0);
+    c = (sc == HISTSUB && HISTSUB != '\0') ? (eChar)HIST : getC(0);
     if (c == (eChar)HIST) {
 	if (alhistp)
 	    return (alhistp);
=============================================================== >8 ==

So the previous example would look something like this:

	> set histchars='\!'
	> echo good
	> !e
	echo good
	> _

What do you think, Christos?

--
nargs <nargs616@yahoo.com>






--
This message has been 'sanitized'.  This means that potentially
dangerous content has been rewritten or removed.  The following
log describes which actions were taken.

Sanitizer (start="1274085533"):
  Split unusually long Date: header.
  SanitizeFile (filename="unnamed.txt", mimetype="text/plain"):
    Match (names="unnamed.txt", rule="2"):
      Enforced policy: accept

  Total modifications so far: 1


Anomy 0.0.0 : Sanitizer.pm
$Id: Sanitizer.pm,v 1.89 2004/09/02 10:14:15 bre Exp $
$Id: gw.conf,v 1.1 2005/06/23 02:18:04 kim Exp $